Move /fill endpoint into separate service so we can horizontally scale the quote endpoint but keep the signing service as a signelton and properly track nonces
