+++ package.json
@@ -88,6 +88,7 @@
         "debug": "^4.1.1",
         "dotenv": "^8.1.0",
         "ethereum-types": "^3.0.0",
+        "ethereumjs-util": "^5.1.1",
         "express": "^4.17.1",
         "express-async-handler": "^1.1.4",
         "http-status-codes": "^1.3.2",
