+++ src/handlers/signer_handlers.ts
@@ -1,19 +1,15 @@
-import { ERC20BridgeSource, SwapQuoterError } from '@0x/asset-swapper';
-import { BigNumber, NULL_ADDRESS } from '@0x/utils';
+import { BigNumber } from '@0x/utils';
 import * as express from 'express';
 import * as HttpStatus from 'http-status-codes';
+import * as _ from 'lodash';
 
-import { CHAIN_ID } from '../config';
-import { DEFAULT_QUOTE_SLIPPAGE_PERCENTAGE } from '../constants';
-import { InternalServerError, RevertAPIError, ValidationError, ValidationErrorCodes } from '../errors';
+import { InternalServerError, RevertAPIError } from '../errors';
 import { logger } from '../logger';
 import { isAPIError, isRevertError } from '../middleware/error_handling';
 import { schemas } from '../schemas/schemas';
 import { SignerService } from '../services/signer_service';
-import { TokenMetadatasForChains } from '../token_metadatas_for_networks';
-import { ChainId, GetSwapQuoteRequestParams } from '../types';
+import { ZeroExTransactionWithoutDomain } from '../types';
 import { schemaUtils } from '../utils/schema_utils';
-import { findTokenAddress, isETHSymbol } from '../utils/token_metadata_utils';
 
 export class SignerHandlers {
     private readonly _signerService: SignerService;
@@ -24,43 +20,18 @@ export class SignerHandlers {
     constructor(signerService: SignerService) {
         this._signerService = signerService;
     }
-    public async whitelistZeroExTransactionAsync(req: express.Request, res: express.Response): Promise<void> {
-        // TODO(fabio): Implement me
-    }
-    public async signZeroExTransactionAsync(req: express.Request, res: express.Response): Promise<void> {
-        // TODO(fabio):
-        // Get 0x txn and signature from body
-        // Verify that affiliate fee is included
-        // Verify that order and 0x txn don't expire in less than 60sec
-        // Verify that 0x txn hash has been whitelisted
+    public async signAndSubmitZeroExTransactionAsync(req: express.Request, res: express.Response): Promise<void> {
+        schemaUtils.validateSchema(req.body, schemas.metaTransactionFillRequestSchema);
 
-        // parse query params
-        const {
-            sellToken,
-            buyToken,
-            sellAmount,
-            buyAmount,
-            takerAddress,
-            slippagePercentage,
-            gasPrice,
-            excludedSources,
-        } = parseGetSwapQuoteRequestParams(req);
-        const isETHSell = isETHSymbol(sellToken);
-        const sellTokenAddress = findTokenAddressOrThrowApiError(sellToken, 'sellToken', CHAIN_ID);
-        const buyTokenAddress = findTokenAddressOrThrowApiError(buyToken, 'buyToken', CHAIN_ID);
+        // parse the request body
+        const { zeroExTransaction, signature } = parsePostTransactionRequestBody(req);
         try {
-            const swapQuote = await this._signerService.calculateSwapQuoteAsync({
-                buyTokenAddress,
-                sellTokenAddress,
-                buyAmount,
-                sellAmount,
-                from: takerAddress,
-                isETHSell,
-                slippagePercentage,
-                gasPrice,
-                excludedSources,
+            const { transactionHash, signedEthereumTransaction } = await this._signerService.signAndSubmitZeroExTransactionAsync(zeroExTransaction, signature);
+            // return the transactionReceipt
+            res.status(HttpStatus.OK).send({
+                transactionHash,
+                signedEthereumTransaction,
             });
-            res.status(HttpStatus.OK).send(swapQuote);
         } catch (e) {
             // If this is already a transformed error then just re-throw
             if (isAPIError(e)) {
@@ -70,74 +41,29 @@ export class SignerHandlers {
             if (isRevertError(e)) {
                 throw new RevertAPIError(e);
             }
-            const errorMessage: string = e.message;
-            // TODO AssetSwapper can throw raw Errors or InsufficientAssetLiquidityError
-            if (
-                errorMessage.startsWith(SwapQuoterError.InsufficientAssetLiquidity) ||
-                errorMessage.startsWith('NO_OPTIMAL_PATH')
-            ) {
-                throw new ValidationError([
-                    {
-                        field: buyAmount ? 'buyAmount' : 'sellAmount',
-                        code: ValidationErrorCodes.ValueOutOfRange,
-                        reason: SwapQuoterError.InsufficientAssetLiquidity,
-                    },
-                ]);
-            }
-            if (errorMessage.startsWith(SwapQuoterError.AssetUnavailable)) {
-                throw new ValidationError([
-                    {
-                        field: 'token',
-                        code: ValidationErrorCodes.ValueOutOfRange,
-                        reason: e.message,
-                    },
-                ]);
-            }
             logger.info('Uncaught error', e);
             throw new InternalServerError(e.message);
         }
     }
-    // tslint:disable-next-line:prefer-function-over-method
-    public async getSwapTokensAsync(_req: express.Request, res: express.Response): Promise<void> {
-        const tokens = TokenMetadatasForChains.map(tm => ({
-            symbol: tm.symbol,
-            address: tm.tokenAddresses[CHAIN_ID],
-            name: tm.name,
-            decimals: tm.decimals,
-        }));
-        const filteredTokens = tokens.filter(t => t.address !== NULL_ADDRESS);
-        res.status(HttpStatus.OK).send({ records: filteredTokens });
-    }
 }
 
-const findTokenAddressOrThrowApiError = (address: string, field: string, chainId: ChainId): string => {
-    try {
-        return findTokenAddress(address, chainId);
-    } catch (e) {
-        throw new ValidationError([
-            {
-                field,
-                code: ValidationErrorCodes.ValueOutOfRange,
-                reason: e.message,
-            },
-        ]);
-    }
-};
+interface PostTransactionRequestBody {
+    zeroExTransaction: ZeroExTransactionWithoutDomain;
+    signature: string;
+}
 
-const parseStringArrForERC20BridgeSources = (excludedSources: string[]): ERC20BridgeSource[] => {
-    return excludedSources.filter((source: string) => source in ERC20BridgeSource) as ERC20BridgeSource[];
-};
-
-const parseGetSwapQuoteRequestParams = (req: express.Request): GetSwapQuoteRequestParams => {
-    // HACK typescript typing does not allow this valid json-schema
-    schemaUtils.validateSchema(req.query, schemas.swapQuoteRequestSchema as any);
-    const takerAddress = req.query.takerAddress;
-    const sellToken = req.query.sellToken;
-    const buyToken = req.query.buyToken;
-    const sellAmount = req.query.sellAmount === undefined ? undefined : new BigNumber(req.query.sellAmount);
-    const buyAmount = req.query.buyAmount === undefined ? undefined : new BigNumber(req.query.buyAmount);
-    const gasPrice = req.query.gasPrice === undefined ? undefined : new BigNumber(req.query.gasPrice);
-    const slippagePercentage = Number.parseFloat(req.query.slippagePercentage || DEFAULT_QUOTE_SLIPPAGE_PERCENTAGE);
-    const excludedSources = req.query.excludedSources === undefined ? undefined : parseStringArrForERC20BridgeSources(req.query.excludedSources.split(','));
-    return { takerAddress, sellToken, buyToken, sellAmount, buyAmount, slippagePercentage, gasPrice, excludedSources };
+const parsePostTransactionRequestBody = (req: any): PostTransactionRequestBody => {
+    const requestBody = req.body;
+    const signature = requestBody.signature;
+    const zeroExTransaction: ZeroExTransactionWithoutDomain = {
+        salt: new BigNumber(requestBody.zeroExTransaction.salt),
+        expirationTimeSeconds: new BigNumber(requestBody.zeroExTransaction.expirationTimeSeconds),
+        gasPrice: new BigNumber(requestBody.zeroExTransaction.gasPrice),
+        signerAddress: requestBody.zeroExTransaction.signerAddress,
+        data: requestBody.zeroExTransaction.data,
+    };
+    return {
+        zeroExTransaction,
+        signature,
+    };
 };
